Check to see if the file/dir exists before deleting it.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@10607 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/helpers/ZipFileHelper.java
@@ -76,8 +76,12 @@ public class ZipFileHelper
     {
         for (File dir : dirsToRemoveList)
         {
-            try {
-                FileUtils.deleteDirectory(dir);
+            try 
+            {
+                if (dir.exists())
+                {
+                    FileUtils.deleteDirectory(dir);
+                }
             } catch (IOException ex)
             {
                 System.err.println(ex);
